Squelched a bunch of warnings
made the syn_table and syn_table_idx extern variables from the
header so clang would stop complaining and added the
declarations to main. When you move those completely into
session.c, remove the extern definition and the declaration in main

+++ options.h
@@ -19,7 +19,7 @@ extern int reverse_dns_flag;
 extern char* live_capture_dev;
 extern char* capture_file;
 extern char* csv_output_file;
-extern int size;
+extern long int size;
 
 void process_options(int argc, char** argv);
